Make internal bounds checks conditionally compiled
We run the test with bounds checks on.
