+++ .travis.yml
@@ -47,7 +47,7 @@ install:
 script:
  - go get -d -t -v ./...
  - go build -x -v ./...
- - go test -x -v ./...
+ - go test -tags bounds -x -v ./...
  - diff <(gofmt -d .) <("")
  - if [[ $TRAVIS_SECURE_ENV_VARS = "true" ]]; then bash test-coverage.sh; fi
